requirements.txt
Package                       Version
----------------------------- ------------
blinker                       1.6.2
cachetools                    5.3.1
certifi                       2023.7.22
charset-normalizer            3.3.0
click                         8.1.7
cloudevents                   1.10.0
db-dtypes                     1.1.1
deprecation                   2.1.0
Flask                         2.3.3
functions-framework           3.4.0
google-analytics-admin        0.20.0
google-api-core               2.12.0
google-auth                   2.23.2
google-auth-oauthlib          1.1.0
google-cloud-bigquery         3.11.4
google-cloud-bigquery-storage 2.22.0
google-cloud-core             2.3.3
google-crc32c                 1.5.0
google-resumable-media        2.6.0
googleapis-common-protos      1.60.0
grpcio                        1.59.0
grpcio-status                 1.59.0
gunicorn                      20.1.0
idna                          3.4
itsdangerous                  2.1.2
Jinja2                        3.1.2
MarkupSafe                    2.1.3
numpy                         1.26.0
oauthlib                      3.2.2
packaging                     23.2
pandas                        2.1.1
pandas-gbq                    0.19.2
pip                           22.3.1
proto-plus                    1.22.3
protobuf                      4.24.3
pyarrow                       13.0.0
pyasn1                        0.5.0
pyasn1-modules                0.3.0
pydata-google-auth            1.8.2
python-dateutil               2.8.2
pytz                          2023.3.post1
requests                      2.31.0
requests-oauthlib             1.3.1
rsa                           4.9
setuptools                    65.5.0
six                           1.16.0
tzdata                        2023.3
urllib3                       2.0.5
watchdog                      3.0.0
Werkzeug                      3.0.0
